requirements.txt
ollama~=0.4.6
asyncio~=3.4.3
pathlib~=1.0.1
typing~=3.7.4.3
piexif~=1.1.3
pillow~=11.1.0
pydantic~=2.10.5
requests~=2.32.3
colorama~=0.4.6